Update abstract to better explain distributed system and WSN
In sentence 4, you refer to WSNs "in large distributed systems" and with "dynamic agents". Neither of these seem very clear from the context of the preceding sentences. Isn't a WSN itself a large distributed system, so what does it mean for it to be in another one? What are agents in the context of WSNs and what does it mean for them to be static vs dynamic?